Split EthAssemblyAdapter from AsmCodeGen
