+++ libyul/backends/evm/AsmCodeGen.h
@@ -16,63 +16,24 @@
 */
 // SPDX-License-Identifier: GPL-3.0
 /**
- * Adaptor between the abstract assembly and eth assembly.
+ * Helper to compile Yul code using libevmasm.
  */
 
 #pragma once
 
 #include <libyul/backends/evm/AbstractAssembly.h>
 #include <libyul/AsmAnalysis.h>
-#include <liblangutil/SourceLocation.h>
-#include <functional>
+#include <liblangutil/EVMVersion.h>
 
 namespace solidity::evmasm
 {
 class Assembly;
-class AssemblyItem;
 }
 
 namespace solidity::yul
 {
 struct Block;
-
-class EthAssemblyAdapter: public AbstractAssembly
-{
-public:
-	explicit EthAssemblyAdapter(evmasm::Assembly& _assembly);
-	void setSourceLocation(langutil::SourceLocation const& _location) override;
-	int stackHeight() const override;
-	void setStackHeight(int height) override;
-	void appendInstruction(evmasm::Instruction _instruction) override;
-	void appendConstant(u256 const& _constant) override;
-	void appendLabel(LabelID _labelId) override;
-	void appendLabelReference(LabelID _labelId) override;
-	size_t newLabelId() override;
-	size_t namedLabel(std::string const& _name) override;
-	void appendLinkerSymbol(std::string const& _linkerSymbol) override;
-	void appendVerbatim(bytes _data, size_t _arguments, size_t _returnVariables) override;
-	void appendJump(int _stackDiffAfter, JumpType _jumpType) override;
-	void appendJumpTo(LabelID _labelId, int _stackDiffAfter, JumpType _jumpType) override;
-	void appendJumpToIf(LabelID _labelId, JumpType _jumpType) override;
-	void appendAssemblySize() override;
-	std::pair<std::shared_ptr<AbstractAssembly>, SubID> createSubAssembly(std::string _name = {}) override;
-	void appendDataOffset(std::vector<SubID> const& _subPath) override;
-	void appendDataSize(std::vector<SubID> const& _subPath) override;
-	SubID appendData(bytes const& _data) override;
-
-	void appendImmutable(std::string const& _identifier) override;
-	void appendImmutableAssignment(std::string const& _identifier) override;
-
-	void markAsInvalid() override;
-
-private:
-	static LabelID assemblyTagToIdentifier(evmasm::AssemblyItem const& _tag);
-	void appendJumpInstruction(evmasm::Instruction _instruction, JumpType _jumpType);
-
-	evmasm::Assembly& m_assembly;
-	std::map<SubID, u256> m_dataHashBySubId;
-	size_t m_nextDataCounter = std::numeric_limits<size_t>::max() / 2;
-};
+struct AsmAnalysisInfo;
 
 class CodeGenerator
 {
@@ -88,5 +49,4 @@ public:
 		bool _optimizeStackAllocation = false
 	);
 };
-
 }
